# static/openapi.yaml
openapi: 3.0.0
info:
  title: Book API
  description: API for managing books and reviews
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8000
paths:
  /books:
    get:
      summary: Get a list of all books in the database
      tags:
        - Books
      responses:
        '200':
          description: A list of books
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: The unique identifier for the book
                      example: 1
                    title:
                      type: string
                      description: The title of the book
                      example: '1984'
                    author:
                      type: string
                      description: The author of the book
                      example: 'George Orwell'
                    genre:
                      type: string
                      description: The genre of the book
                      example: 'Dystopian'
                    year_published:
                      type: integer
                      description: The year the book was published
                      example: 1949
                    summary:
                      type: string
                      description: A brief summary of the book
                      example: 'A novel about a dystopian future under a totalitarian regime.'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Internal server error'
    post:
      summary: Add a new book to the database
      tags:
        - Books
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: The title of the book
                  example: '1984'
                author:
                  type: string
                  description: The author of the book
                  example: 'George Orwell'
                genre:
                  type: string
                  description: The genre of the book
                  example: 'Dystopian'
                year_published:
                  type: integer
                  description: The year the book was published
                  example: 1949
                summary:
                  type: string
                  description: A brief summary of the book
                  example: 'A novel about a dystopian future under a totalitarian regime.'
      responses:
        '201':
          description: Book added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The unique identifier for the book
                    example: 1
                  title:
                    type: string
                    description: The title of the book
                    example: '1984'
                  author:
                    type: string
                    description: The author of the book
                    example: 'George Orwell'
                  genre:
                    type: string
                    description: The genre of the book
                    example: 'Dystopian'
                  year_published:
                    type: integer
                    description: The year the book was published
                    example: 1949
                  summary:
                    type: string
                    description: A brief summary of the book
                    example: 'A novel about a dystopian future under a totalitarian regime.'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Internal server error'
  /books/{id}:
    get:
      summary: Retrieve a specific book by its ID
      tags:
        - Books
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The unique identifier for the book
      responses:
        '200':
          description: The requested book
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The unique identifier for the book
                    example: 1
                  title:
                    type: string
                    description: The title of the book
                    example: '1984'
                  author:
                    type: string
                    description: The author of the book
                    example: 'George Orwell'
                  genre:
                    type: string
                    description: The genre of the book
                    example: 'Dystopian'
                  year_published:
                    type: integer
                    description: The year the book was published
                    example: 1949
                  summary:
                    type: string
                    description: A brief summary of the book
                    example: 'A novel about a dystopian future under a totalitarian regime.'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Book not found'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Internal server error'
    put:
      summary: Update a book's information by its ID
      tags:
        - Books
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The unique identifier for the book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: The title of the book
                  example: '1984'
                author:
                  type: string
                  description: The author of the book
                  example: 'George Orwell'
                genre:
                  type: string
                  description: The genre of the book
                  example: 'Dystopian'
                year_published:
                  type: integer
                  description: The year the book was published
                  example: 1949
                summary:
                  type: string
                  description: A brief summary of the book
                  example: 'A novel about a dystopian future under a totalitarian regime.'
      responses:
        '200':
          description: Book updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The unique identifier for the book
                    example: 1
                  title:
                    type: string
                    description: The title of the book
                    example: '1984'
                  author:
                    type: string
                    description: The author of the book
                    example: 'George Orwell'
                  genre:
                    type: string
                    description: The genre of the book
                    example: 'Dystopian'
                  year_published:
                    type: integer
                    description: The year the book was published
                    example: 1949
                  summary:
                    type: string
                    description: A brief summary of the book
                    example: 'A novel about a dystopian future under a totalitarian regime.'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Book not found'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Internal server error'
    delete:
      summary: Delete a book by its ID
      tags:
        - Books
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The unique identifier for the book
      responses:
        '204':
          description: Book deleted successfully
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Book not found'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Internal server error'
  /books/{id}/reviews:
    get:
      summary: Retrieve all reviews for a book
      tags:
        - Reviews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The unique identifier for the book
      responses:
        '200':
          description: A list of reviews for the book
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: The unique identifier for the review
                      example: 1
                    book_id:
                      type: integer
                      description: The unique identifier for the book
                      example: 1
                    review_text:
                      type: string
                      description: The text of the review
                      example: 'An amazing read!'
                    rating:
                      type: integer
                      description: The rating given in the review
                      example: 5
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Book not found'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Internal server error'
    post:
      summary: Add a review for a book
      tags:
        - Reviews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The unique identifier for the book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                review_text:
                  type: string
                  description: The text of the review
                  example: 'An amazing read!'
                rating:
                  type: integer
                  description: The rating given in the review
                  example: 5
      responses:
        '201':
          description: Review added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The unique identifier for the review
                    example: 1
                  book_id:
                    type: integer
                    description: The unique identifier for the book
                    example: 1
                  review_text:
                    type: string
                    description: The text of the review
                    example: 'An amazing read!'
                  rating:
                    type: integer
                    description: The rating given in the review
                    example: 5
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Book not found'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Internal server error'